+++ apps/rig/config/config.exs
@@ -21,6 +21,8 @@ config :rig, RIG.Subscriptions,
 config :rig, Rig.EventStream.KafkaToFilter,
   # The list of brokers, given by a comma-separated list of host:port items:
   brokers: {:system, :list, "KAFKA_BROKERS", []},
+  serializer: {:system, "KAFKA_SERIALIZER", nil},
+  schema_registry_host: {:system, "KAFKA_SCHEMA_REGISTRY_HOST", "localhost:8081"},
   # The list of topics to consume messages from:
   consumer_topics: {:system, :list, "KAFKA_SOURCE_TOPICS", ["rig"]},
   # If KAFKA_SSL_ENABLED=0, the KAFKA_SSL_* settings are ignored; otherwise, they're required.
@@ -38,6 +40,8 @@ config :rig, Rig.EventStream.KafkaToFilter,
 config :rig, Rig.EventStream.KafkaToHttp,
   # The list of brokers, given by a comma-separated list of host:port items:
   brokers: {:system, :list, "KAFKA_BROKERS", []},
+  serializer: {:system, "KAFKA_SERIALIZER", nil},
+  schema_registry_host: {:system, "KAFKA_SCHEMA_REGISTRY_HOST", "localhost:8081"},
   # The list of topics to consume messages from:
   consumer_topics: {:system, :list, "FIREHOSE_KAFKA_SOURCE_TOPICS", ["rig-firehose"]},
   # If KAFKA_SSL_ENABLED=0, the KAFKA_SSL_* settings are ignored; otherwise, they're required.