+++ apps/rig_kafka/lib/rig_kafka/client.ex
@@ -2,15 +2,18 @@ defmodule RigKafka.Client do
   @moduledoc """
   The Kafka client that holds connections to one or more brokers.
   """
-  require Logger
-  @reconnect_timeout_ms 20_000
-  use GenServer, shutdown: @reconnect_timeout_ms + 5_000
 
   alias RigKafka.Config
+  alias RigKafka.Serializer
 
-  @supervisor RigKafka.DynamicSupervisor
+  require Logger
 
   @type callback :: (any -> :ok | any)
+
+  @reconnect_timeout_ms 20_000
+  @supervisor RigKafka.DynamicSupervisor
+
+  use GenServer, shutdown: @reconnect_timeout_ms + 5_000
 
   defmodule GroupSubscriber do
     @moduledoc """
@@ -18,21 +21,79 @@ defmodule RigKafka.Client do
 
     """
     @behaviour :brod_group_subscriber
+
+    import Record, only: [defrecord: 2, extract: 2]
+
     require Logger
     require Record
-    import Record, only: [defrecord: 2, extract: 2]
+
     defrecord :kafka_message, extract(:kafka_message, from_lib: "brod/include/brod.hrl")
+
+    @type kafka_headers :: list()
 
     @impl :brod_group_subscriber
     def init(_brod_group_id, state) do
       {:ok, state}
     end
 
-    @impl :brod_group_subscriber
-    def handle_message(topic, partition, message, %{callback: callback} = state) do
-      %{offset: offset, value: body} = Enum.into(kafka_message(message), %{})
+    # ---
 
-      case callback.(body) do
+    @spec get_content_type(any(), any()) :: String.t()
+    defp get_content_type(<<0::8, _id::32, _body::binary>>), do: "avro/binary"
+    defp get_content_type(_), do: "application/json"
+
+    defp get_content_type(headers, body) do
+      ce_specversion =
+        case headers do
+          %{specversion: version} ->
+            version
+
+          %{cloudEventsVersion: version} ->
+            version
+
+          _ ->
+            headers
+        end
+
+      case ce_specversion do
+        "0.2" -> Map.get(headers, :contenttype, get_content_type(body))
+        "0.1" -> Map.get(headers, :contentType, get_content_type(body))
+        _ -> get_content_type(body)
+      end
+    end
+
+    # ---
+
+    @impl :brod_group_subscriber
+    def handle_message(
+          topic,
+          partition,
+          message,
+          %{callback: callback, schema_registry_host: schema_registry_host} = state
+        ) do
+      %{offset: offset, value: body, headers: headers} = Enum.into(kafka_message(message), %{})
+      headers_no_prefix = Serializer.remove_prefix(headers)
+
+      content_type = get_content_type(headers_no_prefix, body)
+
+      decoded_body =
+        cond do
+          content_type == "avro/binary" ->
+            data =
+              body
+              |> Serializer.decode_body("avro", schema_registry_host)
+              |> Jason.decode!()
+
+            Map.merge(headers_no_prefix, %{data: data})
+
+          content_type == "application/json" ->
+            body
+
+          true ->
+            body
+        end
+
+      case callback.(decoded_body) do
         :ok ->
           {:ok, :ack, state}
 
@@ -82,12 +143,14 @@ defmodule RigKafka.Client do
 
   # ---
 
-  def produce(%{server_id: server_id}, topic, key, plaintext)
+  def produce(%{server_id: server_id}, topic, schema, key, plaintext)
       when is_binary(topic) and is_binary(key) and is_binary(plaintext) do
-    GenServer.call(server_id, {:produce, topic, key, plaintext})
+    GenServer.call(server_id, {:produce, topic, schema, key, plaintext})
   end
 
   # ---
+
+  @type kafka_headers :: list()
 
   @impl GenServer
   def init(%{config: config} = args) do
@@ -185,7 +248,8 @@ defmodule RigKafka.Client do
          config: %{
            client_id: client_id,
            group_id: group_id,
-           consumer_topics: consumer_topics
+           consumer_topics: consumer_topics,
+           schema_registry_host: schema_registry_host
          },
          callback: callback
        }) do
@@ -199,15 +263,36 @@ defmodule RigKafka.Client do
       group_config,
       consumer_config,
       _callback_module = GroupSubscriber,
-      _callback_init_args = %{callback: callback}
+      _callback_init_args = %{callback: callback, schema_registry_host: schema_registry_host}
     )
   end
 
   # ---
 
   @impl GenServer
-  def handle_call({:produce, topic, key, plaintext}, _from, %{brod_client: brod_client} = state) do
-    result = try_producing_message(brod_client, topic, key, plaintext)
+  def handle_call(
+        {:produce, topic, schema, key, plaintext},
+        _from,
+        %{
+          brod_client: brod_client,
+          config: config
+        } = state
+      ) do
+    %{schema_registry_host: schema_registry_host, serializer: serializer} = config
+
+    result =
+      try_producing_message(
+        %{
+          brod_client: brod_client,
+          schema_registry_host: schema_registry_host,
+          serializer: serializer
+        },
+        topic,
+        schema,
+        key,
+        plaintext
+      )
+
     {:reply, result, state}
   end
 
@@ -227,15 +312,73 @@ defmodule RigKafka.Client do
 
   # ---
 
-  defp try_producing_message(brod_client, topic, key, plaintext, retry_delay_divisor \\ 64)
+  @spec transform_content_type(map()) :: [{String.t(), String.t()}]
+  defp transform_content_type(%{"contenttype" => contenttype}) do
+    [
+      {"ce-contenttype", contenttype}
+    ]
+  end
 
-  defp try_producing_message(brod_client, topic, key, plaintext, retry_delay_divisor) do
+  defp transform_content_type(%{"contentType" => contentType}) do
+    [
+      {"ce-contentType", contentType}
+    ]
+  end
+
+  defp transform_content_type(_), do: []
+
+  # ---
+
+  defp try_producing_message(
+         conf,
+         topic,
+         schema,
+         key,
+         plaintext,
+         retry_delay_divisor \\ 64
+       )
+
+  defp try_producing_message(
+         %{
+           brod_client: brod_client,
+           schema_registry_host: schema_registry_host,
+           serializer: serializer
+         },
+         topic,
+         schema,
+         key,
+         plaintext,
+         retry_delay_divisor
+       ) do
+    {constructed_headers, body} =
+      case Jason.decode(plaintext) do
+        {:ok, plaintext_map} ->
+          case serializer do
+            "avro" ->
+              {data, headers} = Map.pop(plaintext_map, "data", %{})
+              prefixed_headers = Serializer.add_prefix(headers)
+
+              {prefixed_headers,
+               Serializer.encode_body(data, "avro", schema, schema_registry_host)}
+
+            _ ->
+              constructed_headers = transform_content_type(plaintext_map)
+              {constructed_headers, plaintext}
+          end
+
+        {:error, _reason} ->
+          {[], plaintext}
+      end
+
     case :brod.produce_sync(
            brod_client,
            topic,
            &compute_kafka_partition/4,
            key,
-           plaintext
+           %{
+             value: body,
+             headers: constructed_headers
+           }
          ) do
       :ok ->
         :ok
@@ -251,7 +394,7 @@ defmodule RigKafka.Client do
           end)
 
           :timer.sleep(retry_delay_ms)
-          try_producing_message(brod_client, topic, key, plaintext, retry_delay_divisor / 2)
+          try_producing_message(brod_client, topic, key, body, retry_delay_divisor / 2)
         else
           {:error, :leader_not_available}
         end